requirements.txt
edge-tts==6.1.9
gtts==2.5.0
pyttsx3==2.90
requests==2.31.0
Pillow==9.5.0
scipy==1.10.1
moviepy==1.0.3
pyyaml==6.0.1
python-dotenv==1.0.0
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
tqdm==4.66.1
colorama==0.4.6
